Revamp and optimize scaling calculations
