+++ NFImageView/Classes/NFImageView.swift
@@ -168,122 +168,6 @@ open class NFImageView: UIView {
         setNeedsDisplay()
     }
     
-    // Helpers
-    
-    // MARK: - Drawing Helpers
-    
-    fileprivate func contentModeRect(forImage image: UIImage) -> CGRect {
-        let imageSize = CGSize(width: image.size.width * image.scale, height: image.size.height * image.scale)
-        
-        switch contentViewMode {
-            
-        case .originalSize: // image size is retained
-            return caculateContentViewImageFillRect(forSize: imageSize)
-            
-        case .aspectFit: // contents scaled to fit with fixed aspect. remainder is transparent
-            
-            var scaling: CGFloat = 1.0
-            let imageMaxSize = imageSize.width > imageSize.height ? imageSize.width : imageSize.height
-            
-            // get max container bounds
-            if bounds.width > bounds.height {
-                // get max scale of max image bounds to container max bounds
-                let maxScale = bounds.width / imageMaxSize
-                // check if image height would exceed the container bounds
-                if imageSize.height * maxScale > bounds.height {
-                    // use the image height if it exceeds the bounds of the container
-                    scaling = bounds.height / imageSize.height
-                }else{
-                    scaling = maxScale
-                }
-            }else{
-                // get max scale of max image bounds to container max bounds
-                let maxScale = bounds.height / imageMaxSize
-                // check if image height would exceed the container bounds
-                if imageSize.width * maxScale > bounds.width {
-                    // use the image height if it exceeds the bounds of the container
-                    scaling = bounds.width / imageSize.width
-                }else{
-                    scaling = maxScale
-                }
-            }
-            
-            let scaledImageSize = CGSize(width: imageSize.width * scaling, height: imageSize.height * scaling)
-            return caculateContentViewImageFillRect(forSize: scaledImageSize)
-            
-        case .aspectFill: // contents scaled to fill with fixed aspect. some portion of content will be clipped.
-            
-            var scaling: CGFloat = 1.0
-            let widthMargin = bounds.width - imageSize.width
-            let heightMargin = bounds.height - imageSize.height
-            
-            if abs(widthMargin) < abs(heightMargin) {
-                if widthMargin > heightMargin {
-                    scaling = bounds.width / imageSize.width
-                }else{
-                    scaling = bounds.height / imageSize.height
-                }
-            }else{
-                if widthMargin > heightMargin {
-                    scaling = bounds.width / imageSize.width
-                }else{
-                    scaling = bounds.height / imageSize.height
-                }
-                
-                if imageSize.width * scaling < bounds.width || imageSize.height * scaling < bounds.height {
-                    if imageSize.width * scaling < bounds.width {
-                        scaling = bounds.width / imageSize.width
-                    }else{
-                        scaling = bounds.height / imageSize.height
-                    }
-                }
-            }
-            
-            let scaledImageSize = CGSize(width: imageSize.width * scaling, height: imageSize.height * scaling)
-            return caculateContentViewImageFillRect(forSize: scaledImageSize)
-            
-        default: break
-            
-        }
-        
-        // .ScaleToFill
-        return CGRect(x: 0, y: 0, width: bounds.width, height: bounds.height)
-    }
-    
-    fileprivate func caculateContentViewImageFillRect(forSize imageSize: CGSize) -> CGRect {
-        
-        // .Center fill
-        var originX: CGFloat = (bounds.width - imageSize.width) / 2
-        var originY: CGFloat = (bounds.height - imageSize.height) / 2
-        
-        if contentViewFill.contains(.Left) {
-            originX = 0
-        }else if contentViewFill.contains(.Right) {
-            originX = (bounds.width - imageSize.width)
-        }
-        
-        if contentViewFill.contains(.Top) {
-            originY = 0
-        }else if contentViewFill.contains(.Bottom) {
-            originY = (bounds.height - imageSize.height)
-        }
-        
-        let contentRect = CGRect(x: originX, y: originY, width: imageSize.width, height: imageSize.height)
-        return flipContentRect(contentRect)
-    }
-    
-    fileprivate func flipContentRect(_ contentRect: CGRect) -> CGRect {
-        // get transform
-        let transform = CGAffineTransform(translationX: 0.0, y: bounds.height)
-        // flip transform
-        let flipTransform = transform.scaledBy(x: 1.0, y: -1.0)
-        // apply transform to target rect
-        let flipContentRect = contentRect.applying(flipTransform)
-        
-        return flipContentRect
-    }
-    
-    
     // MARK: - Loaders Setup
     
     fileprivate func prepareLoadingIndicator() -> UIActivityIndicatorView {
